refactor: docs structure (#433)
While getting familiar with this repository, I tried to utilize its
documentation, and I noticed what I think are its weaknesses. They all
fall down into several categories:
- missing context (e.g. about why you need to tunnel an application)
- a disorder in structure (e.g. there is a section "Deployment" that
does not feature any information about the actual deployment process,
that information is in a different section)
- too much noise in the top-level README

My idea to solve the problems above is to limit the role of the root
level README to:
- present the outline of what the repository is
- give instructions on how to get started with the repository without
going into details
- redirect to other pages for details

Another thing I noticed is that when you follow the current README and
try to access the applications on all the different ports, most of them
end up "not working". They either display a blank page
(`saleor-app-checkout`) or an error page (missing token in
`checkout-storefront`). You don't know if you did something wrong, or if
this is intentional. I already spoke to @witoszekdev about a potential
solution.

**The best way to review it is to switch to the branch the PR is built
for, and go to `README.md`.**

**Before:**
```
- README.md
- /docs
--  ...
```

**After:**
```
- README.md
- /docs
-- ...
-- development.md ➕
-- /payment ➕
---- index.md ➕
---- mollie.md ➕
---- adien.md ➕
```

- Moved the "payment providers" sections to separate files in a new
folder. The folder is called `payment`, and the files are:
`payment/index.md`, `payment/mollie.md`, and `payment/adien.md`.
- Created a `docs/development.md` file where I moved more detailed
instructions about the development. What's left in the `README.md` is
how to install the repository and boot it up.
- I moved some sections around in the `docs/development.md` to maintain
a clearer structure.
- I described some context around the tunneling situation in the
`saleor-app-checkout` section and the checkout id situation in the
`checkout-storefront` section.
- Did some more proofreading, changed some copies, and added some 🌈
**emojis** 🌈.

- I left a comment in the `saleor-app-checkout` section of the
`docs/development.md` document. There is one command that felt kind of
out of place or perhaps incorrect. Please tell me what to do with it
(either delete it or move it somewhere).
- When I first started working with README, I got quickly overwhelmed
with it. The actions I took upon it were quite drastic, so if I moved
something from the top-level README that you think belongs there, please
let me know.
- I spoke to @witoszekdev about the future of `react-storefront`
documentation and he suggested we may be already at a point where
keeping documentation next to the code is unsustainable. I see two
solutions, one temporary and one permanent:
1. (temporary) Provide the "Table of contents" section in the top level
README that references all the main sections in all the files that
people may be looking for.
2. (permanent) Move the docs to a separate app.

Co-authored-by: Jonatan Witoszek <jonatanwitoszek@gmail.com>

refactor: docs structure (#433)

While getting familiar with this repository, I tried to utilize its
documentation, and I noticed what I think are its weaknesses. They all
fall down into several categories:
- missing context (e.g. about why you need to tunnel an application)
- a disorder in structure (e.g. there is a section "Deployment" that
does not feature any information about the actual deployment process,
that information is in a different section)
- too much noise in the top-level README

My idea to solve the problems above is to limit the role of the root
level README to:
- present the outline of what the repository is
- give instructions on how to get started with the repository without
going into details
- redirect to other pages for details

Another thing I noticed is that when you follow the current README and
try to access the applications on all the different ports, most of them
end up "not working". They either display a blank page
(`saleor-app-checkout`) or an error page (missing token in
`checkout-storefront`). You don't know if you did something wrong, or if
this is intentional. I already spoke to @witoszekdev about a potential
solution.

**The best way to review it is to switch to the branch the PR is built
for, and go to `README.md`.**

**Before:**
```
- README.md
- /docs
--  ...
```

**After:**
```
- README.md
- /docs
-- ...
-- development.md ➕
-- /payment ➕
---- index.md ➕
---- mollie.md ➕
---- adien.md ➕
```

- Moved the "payment providers" sections to separate files in a new
folder. The folder is called `payment`, and the files are:
`payment/index.md`, `payment/mollie.md`, and `payment/adien.md`.
- Created a `docs/development.md` file where I moved more detailed
instructions about the development. What's left in the `README.md` is
how to install the repository and boot it up.
- I moved some sections around in the `docs/development.md` to maintain
a clearer structure.
- I described some context around the tunneling situation in the
`saleor-app-checkout` section and the checkout id situation in the
`checkout-storefront` section.
- Did some more proofreading, changed some copies, and added some 🌈
**emojis** 🌈.

- I left a comment in the `saleor-app-checkout` section of the
`docs/development.md` document. There is one command that felt kind of
out of place or perhaps incorrect. Please tell me what to do with it
(either delete it or move it somewhere).
- When I first started working with README, I got quickly overwhelmed
with it. The actions I took upon it were quite drastic, so if I moved
something from the top-level README that you think belongs there, please
let me know.
- I spoke to @witoszekdev about the future of `react-storefront`
documentation and he suggested we may be already at a point where
keeping documentation next to the code is unsustainable. I see two
solutions, one temporary and one permanent:
1. (temporary) Provide the "Table of contents" section in the top level
README that references all the main sections in all the files that
people may be looking for.
2. (permanent) Move the docs to a separate app.

Co-authored-by: Jonatan Witoszek <jonatanwitoszek@gmail.com>